% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEEBeval-package.R
\docType{package}
\name{DEEBeval-package}
\alias{DEEBeval}
\alias{DEEBeval-package}
\title{DEEBeval: Evaluate DEEB Results}
\description{
Evaluate the submissions to the Differential Equation Estimation Benchmark DEEB.
}
\author{
\strong{Maintainer}: Christof Schötz \email{math@christof-schoetz.de} (\href{https://orcid.org/0000-0003-3528-4544}{ORCID})

}
\keyword{internal}
